Mod password storage + generation for users
The introduced changes modify the way MyBB generates passwords for users
(done when requesting a new password) and how it stores those passwords
in the DB.
The basic structure md5(md5($salt) . md5($pass)) is kept for compability.
But instead of storing this in the DB, that hash is hashed again using
bcrypt with a random 22 character salt and a cost of 12.

This improves data security in case of a data breach.
Note that very weak passwords can still be cracked. Educate your users
and enforce the use of strong passes.